moved AvatarService and LocalService into their own package
