+++ pslogin/src/main/scala/services/avatar/AvatarServiceMessage.scala
@@ -1,0 +1,4 @@
+// Copyright (c) 2017 PSForever
+package services.avatar
+
+final case class AvatarServiceMessage(forChannel : String, actionMessage : AvatarAction.Action)
