+++ pslogin/src/main/scala/services/local/LocalServiceMessage.scala
@@ -1,0 +1,4 @@
+// Copyright (c) 2017 PSForever
+package services.local
+
+final case class LocalServiceMessage(forChannel : String, actionMessage : LocalAction.Action)
